requirements.txt
asgiref==3.7.2
boto==2.49.0
boto3==1.34.59
botocore==1.34.59
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
dj-database-url==2.1.0
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
